# 🎯 Phase F.1: Core file splitting for 7 critical files - Issue #93