--- a/src/config/I18nPerformanceConfig.js
+++ b/src/config/I18nPerformanceConfig.js
@@ -1,6 +1,6 @@
 /**
  * 多言語対応パフォーマンス監視設定（自動生成）
- * 生成日時: 2025-08-03T23:46:19.498Z
+ * 生成日時: 2025-08-04T15:28:53.761Z
  */
 
 export const i18nPerformanceConfig = {
@@ -24,7 +24,7 @@ export const i18nPerformanceConfig = {
     }
   },
   "alerts": {},
-  "generatedAt": "2025-08-03T23:46:19.498Z"
+  "generatedAt": "2025-08-04T15:28:53.761Z"
 };
 
 export default i18nPerformanceConfig;