… .kiro/specs/core-file-splitting-phase-f1/PROJECT_COMPLETION_SUMMARY.md
… .kiro/specs/core-file-splitting-phase-f1/design.md
… .kiro/specs/core-file-splitting-phase-f1/requirements.md
… .kiro/specs/core-file-splitting-phase-f1/tasks.md
… CLAUDE.md
… src/config/FontPreloadConfig.js
… src/core/ChallengeUI_original.js
… src/core/FocusManager.js
… src/core/FocusManager_new.js
… src/core/FocusManager_original.js
… src/core/MotionManager_new.js
… src/core/MotionManager_original.js
… src/core/SettingsManager.js
… src/core/StatisticsDataRecovery.js
… src/core/TimingAdjustmentManager_backup.js
… src/core/TimingAdjustmentManager_new.js
… src/core/challenge/ChallengeInteractionHandler.js
… src/core/challenge/ChallengeUIRenderer.js
… src/core/focus/FocusNavigation.js
… src/core/focus/FocusRingRenderer.js
… src/core/focus/FocusTrapManager.js
… src/core/help/HelpEffectivenessAnalyzer_new.js
… src/core/help/HelpEffectivenessAnalyzer_original.js
… src/core/help/effectiveness/HelpDataAnalyzer.js
… src/core/help/effectiveness/HelpMetricsCollector.js
… src/core/help/effectiveness/HelpReportGenerator.js
… src/core/motion/AnimationController.js
… src/core/motion/MotionConfigManager.js
… src/core/motion/VestibularSafetyManager.js
… src/core/settings/SettingsDataManager.js