## Summary

- ✅ Complete Phase F.1 core file splitting project implementation
- ✅ Split 7 large core files (>2,500 words) using Main Controller Pattern
- ✅ Achieve MCP tool compatibility (all files now <2,500 words)
- ✅ Maintain 100% API compatibility while reducing file sizes by 59%

## Key Changes

### Files Split (7/7 completed)

| File | Original Size | New Size | Reduction | Components |
|------|---------------|----------|-----------|------------|
| SettingsManager.js | 2,812 words | 1,369 words | 51% | 4 components |
| StatisticsDataRecovery.js | 2,772 words | 1,122 words | 60% | 4 components |
| FocusManager.js | 2,765 words | 1,357 words | 51% | 4 components |
| HelpEffectivenessAnalyzer.js | 2,757 words | 963 words | 65% | 4 components |
| MotionManager.js | 2,754 words | 859 words | 69% | 4 components |
| ChallengeUI.js | 2,644 words | 769 words | 71% | 3 components |
| TimingAdjustmentManager.js | 2,535 words | 1,071 words | 58% | 3 components |

**Total Reduction**: 18,439 words → 7,510 words (**59% reduction**)

### Architecture Implementation

- **Main Controller Pattern**: Each file transformed into lightweight orchestrator
- **Dependency Injection**: Sub-components injected into main controllers
- **API Preservation**: All public interfaces maintained unchanged
- **Component Structure**: Organized into specialized sub-components with single responsibility

### New Directory Structure

```
src/core/
├── SettingsManager.js + settings/
├── StatisticsDataRecovery.js + statistics/
├── FocusManager.js + focus/
├── help/HelpEffectivenessAnalyzer.js + help/
├── MotionManager.js + motion/
├── ChallengeUI.js + challenge/
└── TimingAdjustmentManager.js + timing-adjustment/
```

## Validation Results

- ✅ **MCP Tool Compatibility**: All files now under 2,500 words, no token limit errors
- ✅ **Build Process**: Production build successful (3.46s)
- ✅ **API Compatibility**: All public interfaces preserved
- ✅ **Integration Testing**: Core functionality verified
- ✅ **Performance**: No measurable degradation

## Test Plan

- [x] MCP tool compatibility verification (find_symbol works on all files)
- [x] Production build validation
- [x] Integration testing with existing systems
- [x] API compatibility verification
- [x] Performance impact assessment

## Developer Experience Impact

This resolves Issue #70 (MCP token limit problem) by:
- Restoring full MCP tool functionality
- Enabling fast, reliable symbol navigation
- Improving development productivity
- Creating better organized, maintainable code

## Related Issues

- Fixes #93 (Phase F.1 core file splitting)
- Addresses #70 (MCP token limit issue)
- Part of #77 (file size optimization project)

🤖 Generated with [Claude Code](https://claude.ai/code)